libtmux(textframe): Add pytest plugin with hooks and fixtures
why: Auto-register TextFrame assertion hooks and snapshot fixture
for downstream users who install libtmux[textframe].
what:
- Move tests/textframe/plugin.py → src/libtmux/textframe/plugin.py
- Add pytest_assertrepr_compare hook for rich diff output
- Add textframe_snapshot fixture for downstream users
- Export TextFrameExtension from __init__.py
- Simplify tests/textframe/conftest.py (hooks now in plugin)
