+++ src/libtmux/textframe/__init__.py
@@ -3,5 +3,6 @@
 from __future__ import annotations
 
 from libtmux.textframe.core import ContentOverflowError, TextFrame
+from libtmux.textframe.plugin import TextFrameExtension
 
-__all__ = ["ContentOverflowError", "TextFrame"]
+__all__ = ["ContentOverflowError", "TextFrame", "TextFrameExtension"]
